+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+## next release
+    - [fix] autoload section adds to composer.json
+    - [fix] `t.HttpTransport` and `t.UdpTransport`: 'at' symbol makes initialization safely
+    - [fix] `k.g.GelfTarget`: `init` method sets a `MessageValidator` and `Publisher` instances to the DI container; publishing message action extracted to the separately method
+    - [fix] `getPublisher` method: removed second param from `container->get` function 
+    - [fix] `createMessage` method: if $msg is array - message gets a "short" message and "full" message from array keys 
+    - [fix] `createMessage` method: if message instanse of Throwable, `fullMessage` will be a string (stack trace as string)
+    - [new] add public method `getMessageValidator` to `k.g.GraylogTarget`
+    - [new] Created `k.g.p.Publisher` class, which extends `Gelf\Publisher` but have not optional construct params
+
 ## 7.11.2016: 2.0.0 release
     - [new] Implements configuration with using DI cotainer
     - [brake] Method `getTransport` now is NOT protected